requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
strawberry-graphql[fastapi]
python-multipart
cookiecutter
